# Improve accuracy of conversions from floating-point numbers (Fixes #102)