@kimikage 2019-10-28
On the 32-bit versions, some pre-calculated constants do not agree with the dynamically calculated values.
```julia
julia> versioninfo()
Julia Version 1.4.0-DEV.378
Commit dbaa6ff660 (2019-10-28 12:55 UTC)
Platform Info:
  OS: Windows (i686-w64-mingw32)
  CPU: Intel(R) Core(TM) i7-8565U CPU @ 1.80GHz
  WORD_SIZE: 32
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.1 (ORCJIT, skylake)

julia> a(x) = bitstring(x*(typemax(UInt64)/(FixedPointNumbers.rawone(N11f53))));

julia> b(x) = bitstring((x*typemax(UInt64))/(FixedPointNumbers.rawone(N11f53)));

julia> a(1)
"0100000010100000000000000000000000000000000000000000000000000000"

julia> b(1)
"0100000010100000000000000000000000000000000000000000000000000001"

julia> Float64(BigFloat(typemax(UInt64))/BigFloat(FixedPointNumbers.rawone(N11f53)))
2048.0

julia> Float64(typemax(UInt64))/Float64(FixedPointNumbers.rawone(N11f53))
2048.0000000000005

julia> BigFloat(typemax(UInt64))/BigFloat(FixedPointNumbers.rawone(N11f53))
2048.000000000000227262653140769569055940417886493182912978948283864490173983107
```
Is this a problem that should be solved by modifying the tests?