@kimikage 2019-10-29
To avoid impact on the performance, this PR does not change the `rem`. 
https://github.com/JuliaMath/FixedPointNumbers.jl/blob/da3931838dd3a5f03e8ed00542a306d685002e33/src/normed.jl#L67-L70

Therefore, this does not fix the problems such as:
```julia
julia> using Colors

julia> white = parse(RGB{Float32}, "white")
RGB{Float32}(1.0f0,1.0f0,1.0f0)

julia> convert(RGB{N0f32}, white)
RGB{N0f32}(0.0,0.0,0.0)
```