@kimikage 2019-10-31
> Does this have negative consequences for performance?

Unfortunately, my *software* floating-point calculation is ~2.5x slower.
The following are the benchmark results based on the (inverse) `Vec4` version of https://github.com/JuliaMath/FixedPointNumbers.jl/issues/129#issuecomment-544244496

| x86_64              |    before |     after |
|:--------------------|:---------:|:---------:|
| `Float32`->`N0f8`   | 15.999 μs | 14.099 μs |
| `Float32`->`N8f24`  | 20.000 μs | 48.199 μs |
| `Float32`->`N0f32`  | 20.100 μs | 49.100 μs |
| `Float64`->`N0f8`   | 15.999 μs | 13.999 μs |
| `Float64`->`N0f32`  | 16.099 μs | 14.299 μs |
| `Float64`->`N11f53` | 16.099 μs | 14.299 μs |
| `Float64`->`N0f64`  | 40.999 μs | 68.900 μs |

However, we cannot measure the conversion speed of what cannot be converted. As implied above, in applications where the speed is needed, `rem` is used and the conversions with range checking should not be used.