@timholy 2019-10-31
Thanks for posting those numbers. Correctness is infinitely more important than speed, so I would support merging this.