@kimikage 2019-11-01
For the lines based on the current source code, I reverted the comparison manner. (cf. https://github.com/JuliaMath/FixedPointNumbers.jl/pull/131#discussion_r340806823)

In the new conversion method, I used the `a <= x <= b` style. This decision is based on [my policy](https://github.com/JuliaMath/FixedPointNumbers.jl/pull/131#discussion_r341069209) and the rough benchmark. Honestly, the reason is that it is shorter in source codes. :smiley:
Although the inconsistency makes me somewhat itchy, I think it will also make more people notice this comparison issue.