@kimikage 2019-11-01
I am ready to rebase and merge this PR, but not yet ready for other issues. When When do you bump the version number up?