@timholy 2019-11-01
Version numbers are cheap: any time it's potentially breaking, we should bump the minor version number. If the other fixes will follow quickly I'd say let's consolidate the churn and wait to tag a release until it's all done. If not, let's get a release out for the fixes we have.

As the community has moved to putting [upper version bounds on packages](https://discourse.julialang.org/t/announcement-automatic-merging-for-the-general-registry/29961), it's also more likely that breaking changes won't break things for users.